changes related to tilemap editor

+++ SFMLSuperGame/src/Core/AnimationSequencer.cpp
@@ -15,14 +15,17 @@ int AnimationSequencer::loadAnimationFrames(int animationType, const std::string
     // Resize the vector in the map to store the number of frames
     animationFrames[animationType].resize(totalFrames);
 
-    // Loop to extract each frame from the spritesheet using sf::IntRect
+    // Extract each frame from the spritesheet using sf::IntRect
     for (int i = 0; i < totalFrames; ++i)
     {
-        sf::IntRect frameRect(i * frameWidth, 0, frameWidth, frameHeight);  // Define the frame boundaries
-
+        // Define the frame boundaries
+        sf::IntRect frameRect(i * frameWidth, 0, frameWidth, frameHeight);  
+        
         std::cout << "Loaded frame " << i << " for animation type " << animationType << std::endl;
+        
         // Set the texture and texture rectangle for each sprite
         animationFrames[animationType][i].setTexture(*texture);
+        // 
         animationFrames[animationType][i].setTextureRect(frameRect);
     }
 
@@ -33,7 +36,7 @@ int AnimationSequencer::loadAnimationFrames(int animationType, const std::string
 sf::Sprite& AnimationSequencer::getCurrentSprite(int animationType, int currentFrame)
 {
     int totalFrames;  // Get the total number of frames
-    totalFrames = animationFrames[animationType].size();
+    totalFrames = static_cast<int>(animationFrames[animationType].size());
     int frameIndex = currentFrame % totalFrames;  // Ensure we loop through the frames correctly
 
     return animationFrames[animationType][frameIndex];  // Return the correct frame
